Fix wrong data in  GHSA-c32m-27pj-4xcj.json
The fixed version for module org.xwiki.platform:xwiki-platform-security-requiredrights-default and affect version 16.5.0RC1 wasn't following what's indicated in https://github.com/xwiki/xwiki-platform/security/advisories/GHSA-c32m-27pj-4xcj and for some reason is inconsistent with the data provided in other modules of the same file. 
As a result, this was creating for us a false positive in our automated scanner for CVE since it found the CVE for versions after 16.10.3.